project/requirements.txt
flask
openai
torch
ftfy
regex
tqdm
Pillow
streamlit
transformers